Various improvements
