+++ src/test/java/com/github/skjolber/unzip/UnivocityZipFileTest.java
@@ -4,29 +4,33 @@ import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-public class ZipFileTest {
+public class UnivocityZipFileTest {
 
 	@Test
 	public void testMultiThread() throws IOException {
-		FileEntryHandler handler = new ChunkSplitterFileEntryHandler(32 * 1024 * 1024, new TestUnivocityCsvFileEntryHandler(new NoopUnivocityCsvLineHandlerFactory()));
+		FileEntryHandler handler = new ChunkSplitterFileEntryHandler(2 * 1024 * 1024, new TestUnivocityCsvFileEntryHandler(new NoopUnivocityCsvLineHandlerFactory()));
 		
 		File file = new File("./src/test/resources/static/rb_norway-aggregated-gtfs.zip");
 		
+		List<String> list = Arrays.asList("trips.txt");
+
 		System.gc();
 		long time = System.currentTimeMillis();
 		int count = Math.max(8, Runtime.getRuntime().availableProcessors());
 		
 		ZipFileEngine engine = new ZipFileEngine(handler, count);
 		try {
-			assertTrue(engine.handle(new FileZipFileFactory(file)));
+			assertTrue(engine.handle(new FileZipFileFactory(file), list));
 		} finally {
 			engine.close();
 		}
 		
-		System.out.println("Used " + (System.currentTimeMillis() - time) + " millis for file size " + (file.length() / (1024*1024)) + " MB for " + count + " threads");
+		System.out.println("Used " + (System.currentTimeMillis() - time) + " millis for file size " + (file.length() / (1024*1024)) + " MB for " + count + " threads using Univocity parser");
 	}
 
 	@Test
@@ -34,18 +38,20 @@ public class ZipFileTest {
 		FileEntryHandler handler = new TestUnivocityCsvFileEntryHandler(new NoopUnivocityCsvLineHandlerFactory());
 		
 		File file = new File("./src/test/resources/static/rb_norway-aggregated-gtfs.zip");
-		
+
+		List<String> list = Arrays.asList("trips.txt");
+
 		System.gc();
 		long time = System.currentTimeMillis();
 		
 		ZipFileEngine engine = new ZipFileEngine(handler, 1);
 		try {
-			assertTrue(engine.handle(new FileZipFileFactory(file)));
+			assertTrue(engine.handle(new FileZipFileFactory(file), list));
 		} finally {
 			engine.close();
 		}
 		
-		System.out.println("Used " + (System.currentTimeMillis() - time) + " millis for file size " + (file.length() / (1024*1024)) + " MB for " + 1 + " threads");
+		System.out.println("Used " + (System.currentTimeMillis() - time) + " millis for file size " + (file.length() / (1024*1024)) + " MB for " + 1 + " threads using Univocity parser");
 	}
 
 }
